Change patron to confirmedContry + refactor
