+++ backend/app/config.py
@@ -44,7 +44,7 @@ class Settings(BaseSettings):
         "IT",
         "LV",
         "MX",
-        # "NL",
+        "NL",
         "NZ",
         "NO",
         "PL",
