+++ frontend/src/routes/__layout.svelte
@@ -6,7 +6,7 @@
   import BackToTopButton from "../components/back_to_top_button.svelte"
   import CookieDisclaimer from "../components/cookie_disclaimer.svelte"
   import MediaQuery from "svelte-media-query"
-  import CountryLocator from "../components/countryLocator.svelte"
+  import CountryLocator from "../components/country_locator.svelte"
 </script>
 
 <div class="flex flex-col h-screen">
