+++ frontend/src/stores/country.js
@@ -2,3 +2,8 @@ import { persist, localStorage } from "@macfja/svelte-persistent-store"
 import { writable } from "svelte/store"
 
 export const currentCountry = persist(writable("DK"), localStorage(), "currentCountry")
+export const confirmedCountry = persist(
+  writable(false),
+  localStorage(),
+  "confirmedCountry"
+)
