--- frontend/src/stores/patron.js
@@ -1,4 +1,0 @@
-import { persist, localStorage } from "@macfja/svelte-persistent-store"
-import { writable } from "svelte/store"
-
-export const patron = persist(writable(false), localStorage(), "patron")
